+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Treasury page shell
 - Bounties page shell
 - Tech comm page shell
+- Params info on recent events
 
 ## [0.1.0-rc.2] - 2021-10-15
 #### Added